+++ kedro_mlflow/io/metrics/mlflow_metric_dataset.py
@@ -18,13 +18,14 @@ class MlflowMetricDataset(MlflowAbstractMetricDataset):
         run_id: str = None,
         load_args: Dict[str, Any] = None,
         save_args: Dict[str, Any] = None,
+        metadata: Dict[str, Any] | None = None,
     ):
         """Initialise MlflowMetricDataset.
         Args:
             run_id (str): The ID of the mlflow run where the metric should be logged
         """
 
-        super().__init__(key, run_id, load_args, save_args)
+        super().__init__(key, run_id, load_args, save_args, metadata)
 
         # We add an extra argument mode="overwrite" / "append" to enable logging update an existing metric
         # this is not an offical mlflow argument for log_metric, so we separate it from the others