+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Fixed
 
--   :bug: Reopen the mlflow run before each node to bypass mlflow thread safety and ensure all tracking is done within the same run_id ([#623](https://github.com/Galileo-Galilei/kedro-mlflow/issues/623), [#624](https://github.com/Galileo-Galilei/kedro-mlflow/issues/623))
+-   :bug: Reopen the mlflow run before each node to bypass mlflow thread safety and ensure all tracking is done within the same run_id ([#623](https://github.com/Galileo-Galilei/kedro-mlflow/issues/623), [#624](https://github.com/Galileo-Galilei/kedro-mlflow/issues/624))
 
 ## [0.14.2] - 2025-02-16